# api/internal/api/zz_generated.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for ProductStatus.
const (
	Available ProductStatus = "available"
	Hidden    ProductStatus = "hidden"
	SoldOut   ProductStatus = "sold_out"
)

// Defines values for ProductType.
const (
	GiftBox    ProductType = "gift_box"
	SingleCake ProductType = "single_cake"
)

// AdminMe defines model for AdminMe.
type AdminMe struct {
	Email string             `json:"email"`
	Id    openapi_types.UUID `json:"id"`
	Name  *string            `json:"name"`
}

// Error defines model for Error.
type Error struct {
	Error string `json:"error"`
}

// Health defines model for Health.
type Health struct {
	Status string `json:"status"`
}

// ImageUploadResult defines model for ImageUploadResult.
type ImageUploadResult struct {
	// ImageUrl Đường dẫn public của ảnh, vd "/uploads/<uuid>.png"
	ImageUrl string `json:"image_url"`
}

// LeadCreated defines model for LeadCreated.
type LeadCreated struct {
	Id openapi_types.UUID `json:"id"`
}

// LeadInput defines model for LeadInput.
type LeadInput struct {
	// Message Lời nhắn
	Message *string `json:"message"`

	// Name Tên khách hàng (bắt buộc)
	Name string `json:"name"`

	// Phone Số điện thoại Việt Nam (bắt buộc)
	Phone string `json:"phone"`

	// ProductInterest Sản phẩm khách quan tâm
	ProductInterest *string `json:"product_interest"`
}

// LoginInput defines model for LoginInput.
type LoginInput struct {
	// Email Email admin
	Email string `json:"email"`

	// Password Mật khẩu (không log, không lưu plaintext)
	Password string `json:"password"`
}

// LoginResult defines model for LoginResult.
type LoginResult struct {
	Ok bool `json:"ok"`
}

// Product defines model for Product.
type Product struct {
	// Badge Nhãn marketing (vd "Bán chạy", "Mới"), null nếu không có
	Badge *string `json:"badge"`

	// CompareAtPrice Giá gốc để so sánh (VND). Nếu > price → hiện giá gạch + % giảm. Null nếu không khuyến mãi.
	CompareAtPrice *int64             `json:"compare_at_price"`
	Description    *string            `json:"description"`
	DisplayOrder   int                `json:"display_order"`
	Id             openapi_types.UUID `json:"id"`
	ImageUrl       *string            `json:"image_url"`
	Name           string             `json:"name"`

	// Price Giá VND, số nguyên
	Price  int64         `json:"price"`
	Slug   string        `json:"slug"`
	Status ProductStatus `json:"status"`

	// Subtitle Nhãn phân loại nhỏ in hoa trên tên (vd "Hộp thiếc cao cấp"), null nếu không có
	Subtitle *string     `json:"subtitle"`
	Type     ProductType `json:"type"`
}

// ProductStatus defines model for Product.Status.
type ProductStatus string

// ProductType defines model for Product.Type.
type ProductType string

// ProductInput defines model for ProductInput.
type ProductInput struct {
	Badge          *string `json:"badge"`
	CompareAtPrice *int64  `json:"compare_at_price"`
	Description    *string `json:"description"`

	// DisplayOrder Thứ tự hiển thị (mặc định 0 nếu bỏ trống)
	DisplayOrder *int    `json:"display_order"`
	ImageUrl     *string `json:"image_url"`

	// Name Tên sản phẩm (bắt buộc, không rỗng)
	Name string `json:"name"`

	// Price Giá VND, số nguyên ≥ 0
	Price int64 `json:"price"`

	// Slug Định danh URL (duy nhất). Trùng → 409.
	Slug string `json:"slug"`

	// Status Trạng thái. Hợp lệ- available | sold_out | hidden
	Status   string  `json:"status"`
	Subtitle *string `json:"subtitle"`

	// Type Loại sản phẩm. Hợp lệ- gift_box | single_cake
	Type string `json:"type"`
}

// UploadProductImageMultipartBody defines parameters for UploadProductImage.
type UploadProductImageMultipartBody struct {
	File openapi_types.File `json:"file"`
}

// CreateProductJSONRequestBody defines body for CreateProduct for application/json ContentType.
type CreateProductJSONRequestBody = ProductInput

// UpdateProductJSONRequestBody defines body for UpdateProduct for application/json ContentType.
type UpdateProductJSONRequestBody = ProductInput

// UploadProductImageMultipartRequestBody defines body for UploadProductImage for multipart/form-data ContentType.
type UploadProductImageMultipartRequestBody UploadProductImageMultipartBody

// LoginJSONRequestBody defines body for Login for application/json ContentType.
type LoginJSONRequestBody = LoginInput

// CreateLeadJSONRequestBody defines body for CreateLead for application/json ContentType.
type CreateLeadJSONRequestBody = LeadInput

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Thông tin admin đang đăng nhập
	// (GET /admin/me)
	GetAdminMe(w http.ResponseWriter, r *http.Request)
	// Danh sách sản phẩm cho admin (gồm cả hidden)
	// (GET /admin/products)
	ListAdminProducts(w http.ResponseWriter, r *http.Request)
	// Tạo sản phẩm mới
	// (POST /admin/products)
	CreateProduct(w http.ResponseWriter, r *http.Request)
	// Xóa (ẩn) sản phẩm
	// (DELETE /admin/products/{id})
	DeleteProduct(w http.ResponseWriter, r *http.Request, id openapi_types.UUID)
	// Cập nhật sản phẩm
	// (PUT /admin/products/{id})
	UpdateProduct(w http.ResponseWriter, r *http.Request, id openapi_types.UUID)
	// Upload ảnh sản phẩm
	// (POST /admin/products/{id}/image)
	UploadProductImage(w http.ResponseWriter, r *http.Request, id openapi_types.UUID)
	// Đăng nhập admin
	// (POST /auth/login)
	Login(w http.ResponseWriter, r *http.Request)
	// Đăng xuất admin
	// (POST /auth/logout)
	Logout(w http.ResponseWriter, r *http.Request)
	// Health check
	// (GET /healthz)
	GetHealthz(w http.ResponseWriter, r *http.Request)
	// Tạo lead từ form liên hệ
	// (POST /leads)
	CreateLead(w http.ResponseWriter, r *http.Request)
	// Danh sách sản phẩm public
	// (GET /products)
	ListProducts(w http.ResponseWriter, r *http.Request)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Thông tin admin đang đăng nhập
// (GET /admin/me)
func (_ Unimplemented) GetAdminMe(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Danh sách sản phẩm cho admin (gồm cả hidden)
// (GET /admin/products)
func (_ Unimplemented) ListAdminProducts(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Tạo sản phẩm mới
// (POST /admin/products)
func (_ Unimplemented) CreateProduct(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Xóa (ẩn) sản phẩm
// (DELETE /admin/products/{id})
func (_ Unimplemented) DeleteProduct(w http.ResponseWriter, r *http.Request, id openapi_types.UUID) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Cập nhật sản phẩm
// (PUT /admin/products/{id})
func (_ Unimplemented) UpdateProduct(w http.ResponseWriter, r *http.Request, id openapi_types.UUID) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Upload ảnh sản phẩm
// (POST /admin/products/{id}/image)
func (_ Unimplemented) UploadProductImage(w http.ResponseWriter, r *http.Request, id openapi_types.UUID) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Đăng nhập admin
// (POST /auth/login)
func (_ Unimplemented) Login(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Đăng xuất admin
// (POST /auth/logout)
func (_ Unimplemented) Logout(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Health check
// (GET /healthz)
func (_ Unimplemented) GetHealthz(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Tạo lead từ form liên hệ
// (POST /leads)
func (_ Unimplemented) CreateLead(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Danh sách sản phẩm public
// (GET /products)
func (_ Unimplemented) ListProducts(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetAdminMe operation middleware
func (siw *ServerInterfaceWrapper) GetAdminMe(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetAdminMe(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ListAdminProducts operation middleware
func (siw *ServerInterfaceWrapper) ListAdminProducts(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListAdminProducts(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CreateProduct operation middleware
func (siw *ServerInterfaceWrapper) CreateProduct(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateProduct(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// DeleteProduct operation middleware
func (siw *ServerInterfaceWrapper) DeleteProduct(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "id" -------------
	var id openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "id", chi.URLParam(r, "id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.DeleteProduct(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// UpdateProduct operation middleware
func (siw *ServerInterfaceWrapper) UpdateProduct(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "id" -------------
	var id openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "id", chi.URLParam(r, "id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.UpdateProduct(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// UploadProductImage operation middleware
func (siw *ServerInterfaceWrapper) UploadProductImage(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "id" -------------
	var id openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "id", chi.URLParam(r, "id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.UploadProductImage(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// Login operation middleware
func (siw *ServerInterfaceWrapper) Login(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.Login(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// Logout operation middleware
func (siw *ServerInterfaceWrapper) Logout(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.Logout(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetHealthz operation middleware
func (siw *ServerInterfaceWrapper) GetHealthz(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetHealthz(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CreateLead operation middleware
func (siw *ServerInterfaceWrapper) CreateLead(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateLead(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ListProducts operation middleware
func (siw *ServerInterfaceWrapper) ListProducts(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListProducts(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/admin/me", wrapper.GetAdminMe)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/admin/products", wrapper.ListAdminProducts)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/admin/products", wrapper.CreateProduct)
	})
	r.Group(func(r chi.Router) {
		r.Delete(options.BaseURL+"/admin/products/{id}", wrapper.DeleteProduct)
	})
	r.Group(func(r chi.Router) {
		r.Put(options.BaseURL+"/admin/products/{id}", wrapper.UpdateProduct)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/admin/products/{id}/image", wrapper.UploadProductImage)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/auth/login", wrapper.Login)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/auth/logout", wrapper.Logout)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/healthz", wrapper.GetHealthz)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/leads", wrapper.CreateLead)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/products", wrapper.ListProducts)
	})

	return r
}